@@ -1,47 +1,76 @@
 {
   "name": "action-ktlint",
-  "version": "4.0.0",
-  "private": false,
   "description": "GitHub Actions for ktlint",
-  "main": "lib/main.js",
-  "scripts": {
-    "build": "tsc",
-    "format": "prettier --write **/*.ts",
-    "format-check": "prettier --check **/*.ts",
-    "lint": "eslint src/**/*.ts",
-    "package": "ncc build --source-map --license licenses.txt",
-    "test": "jest",
-    "all": "npm run build && npm run format && npm run lint && npm run package && npm test"
-  },
+  "version": "4.0.0",
+  "author": "yutailang0119",
+  "type": "module",
+  "private": false,
+  "homepage": "https://github.com/yutailang0119/action-ktlint",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/yutailang0119/action-ktlint.git"
+  },
+  "bugs": {
+    "url": "https://github.com/yutailang0119/action-ktlint/issues"
   },
   "keywords": [
     "actions",
     "ktlint",
     "android"
   ],
-  "author": "yutailang0119",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "engines": {
+    "node": ">=20"
+  },
+  "scripts": {
+    "bundle": "npm run format:write && npm run package",
+    "ci-test": "NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 npx jest",
+    "coverage": "npx make-coverage-badge --output-path ./badges/coverage.svg",
+    "format:write": "npx prettier --write .",
+    "format:check": "npx prettier --check .",
+    "lint": "npx eslint .",
+    "local-action": "npx @github/local-action . src/main.ts .env",
+    "package": "npx rollup --config rollup.config.ts --configPlugin @rollup/plugin-typescript",
+    "package:watch": "npm run package -- --watch",
+    "test": "NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 npx jest",
+    "all": "npm run format:write && npm run lint && npm run test && npm run coverage && npm run package"
+  },
   "license": "MIT",
   "dependencies": {
-    "@actions/core": "^1.10.1",
-    "@actions/glob": "^0.4.0",
+    "@actions/core": "^1.11.1",
+    "@actions/glob": "^0.5.0",
     "xml2js": "^0.6.2"
   },
   "devDependencies": {
-    "@types/node": "^20.12.4",
+    "@eslint/compat": "^1.2.8",
+    "@github/local-action": "^3.1.4",
+    "@jest/globals": "^29.7.0",
+    "@rollup/plugin-commonjs": "^28.0.1",
+    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@rollup/plugin-typescript": "^12.1.1",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^20.17.30",
     "@types/xml2js": "^0.4.14",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
-    "@vercel/ncc": "^0.38.1",
-    "eslint": "^8.57.0",
-    "eslint-plugin-github": "^4.8.0",
-    "eslint-plugin-jest": "^27.9.0",
-    "jest": "^27.5.1",
-    "js-yaml": "^4.1.0",
-    "prettier": "^2.8.8",
-    "ts-jest": "^27.1.3",
-    "typescript": "^4.9.5"
+    "@typescript-eslint/eslint-plugin": "^8.29.1",
+    "@typescript-eslint/parser": "^8.29.1",
+    "eslint": "^9.24.0",
+    "eslint-config-prettier": "^10.1.2",
+    "eslint-import-resolver-typescript": "^4.3.2",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jest": "^28.11.0",
+    "eslint-plugin-prettier": "^5.2.6",
+    "jest": "^29.7.0",
+    "make-coverage-badge": "^1.2.0",
+    "prettier": "^3.5.3",
+    "prettier-eslint": "^16.3.0",
+    "rollup": "^4.40.0",
+    "ts-jest": "^29.3.2",
+    "ts-jest-resolver": "^2.0.1",
+    "typescript": "^5.8.3"
+  },
+  "optionalDependencies": {
+    "@rollup/rollup-linux-x64-gnu": "*"
   }
 }
